requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
easyocr==1.6.2
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
imageio==2.25.0
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
networkx==3.0
ninja==1.11.1
numpy==1.24.2
opencv-python==4.7.0.68
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyclipper==1.3.0.post4
Pygments==2.14.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tifffile==2023.2.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
